+++ packages/ilmomasiina-backend/src/mail/config.ts
@@ -5,6 +5,20 @@ import mailgun from "nodemailer-mailgun-transport";
 import config from "../config";
 
 const mailTransporter: Transporter = (() => {
+  if (process.env.NODE_ENV === "test") {
+    return nodemailer.createTransport({
+      name: "test mail service",
+      version: "0",
+      send(mail, callback) {
+        // Ignore emails in test environment
+        const { message } = mail;
+        const envelope = message.getEnvelope();
+        const messageId = message.messageId();
+        setImmediate(() => callback(null, { envelope, messageId } as any));
+      },
+    });
+  }
+
   if (config.mailgunApiKey) {
     if (!config.mailgunDomain) {
       throw new Error("Invalid email config: MAILGUN_DOMAIN must be set with MAILGUN_API_KEY.");