+++ packages/ilmomasiina-frontend/src/containers/requireAuth.tsx
@@ -1,24 +1,32 @@
 import React, { ComponentType, useEffect } from "react";
 
-import { loginExpired, redirectToLogin } from "../modules/auth/actions";
-import { useTypedDispatch, useTypedSelector } from "../store/reducers";
+import { useTranslation } from "react-i18next";
+import { useNavigate } from "react-router";
+
+import { loginToast } from "../modules/auth";
+import useStore from "../modules/store";
+import paths from "../paths";
 
 export default function requireAuth<P extends {}>(WrappedComponent: ComponentType<P>) {
   const RequireAuth = (props: P) => {
-    const dispatch = useTypedDispatch();
+    const navigate = useNavigate();
+    const { t } = useTranslation();
 
-    const { accessToken } = useTypedSelector((state) => state.auth);
+    const { accessToken, resetAuth } = useStore((state) => state.auth);
 
     const expired = accessToken && accessToken.expiresAt < Date.now();
     const needLogin = expired || !accessToken;
 
     useEffect(() => {
       if (expired) {
-        dispatch(loginExpired());
+        resetAuth();
+        loginToast("error", t("auth.loginExpired"), 10000);
+        navigate(paths.adminLogin);
       } else if (needLogin) {
-        dispatch(redirectToLogin());
+        resetAuth();
+        navigate(paths.adminLogin);
       }
-    }, [needLogin, expired, dispatch]);
+    }, [needLogin, expired, resetAuth, navigate, t]);
 
     return needLogin ? null : <WrappedComponent {...props} />;
   };