+++ packages/ilmomasiina-frontend/src/routes/AdminEventsList/AdminEventListItem.tsx
@@ -9,10 +9,10 @@ import { toast } from "react-toastify";
 import { ApiError } from "@tietokilta/ilmomasiina-client";
 import { errorDesc } from "@tietokilta/ilmomasiina-client/dist/utils/errorMessage";
 import type { AdminEventListItem as AdminEventListItemSchema } from "@tietokilta/ilmomasiina-models";
+import LinkButton from "../../components/LinkButton";
 import type { TKey } from "../../i18n";
-import { deleteEvent, getAdminEvents } from "../../modules/adminEvents/actions";
+import useStore from "../../modules/store";
 import paths from "../../paths";
-import { useTypedDispatch } from "../../store/reducers";
 import { useEventDateFormatter } from "../../utils/dateFormat";
 import { isEventHiddenFromUsersDueToAge } from "../../utils/eventState";
 
@@ -21,7 +21,7 @@ type Props = {
 };
 
 const AdminEventListItem = ({ event }: Props) => {
-  const dispatch = useTypedDispatch();
+  const { deleteEvent, getAdminEvents } = useStore((state) => state.adminEvents);
 
   const { id, title, slug, date, draft, listed, quotas } = event;
 
@@ -34,13 +34,13 @@ const AdminEventListItem = ({ event }: Props) => {
     const confirmed = window.confirm(t("adminEvents.action.delete.confirm"));
     if (confirmed) {
       try {
-        await dispatch(deleteEvent(id));
+        await deleteEvent(id);
       } catch (err) {
         toast.error(t(errorDesc<TKey>(err as ApiError, "adminEvents.action.delete.error")), {
           autoClose: 2000,
         });
       } finally {
-        dispatch(getAdminEvents());
+        getAdminEvents();
       }
     }
   }
@@ -65,12 +65,12 @@ const AdminEventListItem = ({ event }: Props) => {
       <td>{status}</td>
       <td>{sumBy(quotas, "signupCount")}</td>
       <td>
-        <Button as={Link} variant="primary" size="sm" className="mr-1" to={paths.adminEditEvent(id)}>
+        <LinkButton variant="primary" size="sm" className="me-1" to={paths.adminEditEvent(id)}>
           {t("adminEvents.action.edit")}
-        </Button>
-        <Button as={Link} variant="primary" size="sm" className="mr-1" to={paths.adminCopyEvent(id)}>
+        </LinkButton>
+        <LinkButton variant="primary" size="sm" className="me-1" to={paths.adminCopyEvent(id)}>
           {t("adminEvents.action.copy")}
-        </Button>
+        </LinkButton>
         <Button variant="danger" size="sm" onClick={onDelete}>
           {t("adminEvents.action.delete")}
         </Button>