+++ packages/ilmomasiina-frontend/src/routes/AuditLog/AuditLogFilter.tsx
@@ -4,8 +4,7 @@ import { Form } from "react-bootstrap";
 import { useTranslation } from "react-i18next";
 
 import type { AuditLoqQuery } from "@tietokilta/ilmomasiina-models";
-import { setAuditLogQueryField } from "../../modules/auditLog/actions";
-import { useTypedDispatch } from "../../store/reducers";
+import useStore from "../../modules/store";
 import useThrottled from "../../utils/useThrottled";
 
 const UPDATE_DELAY = 500;
@@ -15,11 +14,11 @@ type Props = Omit<React.ComponentProps<typeof Form.Control>, "name" | "value" | 
 };
 
 const AuditLogFilter = ({ name, ...props }: Props) => {
-  const dispatch = useTypedDispatch();
+  const setAuditLogQueryField = useStore((state) => state.auditLog.setAuditLogQueryField);
   const { t } = useTranslation();
 
   const onChange = useThrottled((e: React.ChangeEvent<HTMLInputElement>) => {
-    dispatch(setAuditLogQueryField(name, e.target.value));
+    setAuditLogQueryField(name, e.target.value);
   }, UPDATE_DELAY);
 
   return <Form.Control type="text" name={name} onChange={onChange} placeholder={t("auditLog.filter")} {...props} />;