+++ packages/ilmomasiina-frontend/src/routes/Editor/components/EditForm.tsx
@@ -5,16 +5,15 @@ import arrayMutators from "final-form-arrays";
 import { Form as BsForm } from "react-bootstrap";
 import { Form, FormRenderProps } from "react-final-form";
 import { useTranslation } from "react-i18next";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router";
 import { toast } from "react-toastify";
 
 import { ApiError, errorDesc } from "@tietokilta/ilmomasiina-client";
 import type { TKey } from "../../../i18n";
-import { publishEventUpdate, publishNewEvent } from "../../../modules/editor/actions";
 import { selectFormData, serverEventToEditor } from "../../../modules/editor/selectors";
 import type { EditorEvent } from "../../../modules/editor/types";
+import useStore from "../../../modules/store";
 import paths from "../../../paths";
-import { useTypedDispatch, useTypedSelector } from "../../../store/reducers";
 import convertZodError from "../../../utils/convertZodError";
 import useEvent from "../../../utils/useEvent";
 import editorSchema from "../schema";
@@ -35,7 +34,7 @@ const EditFormBody = ({ form }: FormRenderProps<EditorEvent>) => {
   const [activeTab, setActiveTab] = useState<EditorTab>(EditorTab.BASIC_DETAILS);
   const { t } = useTranslation();
 
-  const isDraft = useTypedSelector((state) => state.editor.event?.draft || state.editor.isNew);
+  const isDraft = useStore((state) => state.editor.event?.draft || state.editor.isNew);
 
   const doSubmit = useEvent(() => {
     // submit() returns a Promise if validation succeeds.
@@ -91,7 +90,7 @@ const mutators = {
   // The default remove() mutator deletes the entire field if the array becomes empty.
   // This is written based on final-form-arrays's source code to undo that stupidity.
   // See https://github.com/final-form/final-form-arrays/blob/master/src/remove.js
-  remove: ([name, index]: [string, number], state, tools) => {
+  remove: ([name, index], state, tools) => {
     const returnValue = arrayMutators.remove([name, index], state, tools);
     tools.changeValue(state, name, (value) => value ?? []);
     return returnValue;
@@ -104,23 +103,20 @@ function validate(event: EditorEvent) {
 }
 
 const EditForm = () => {
-  const initialValues = useTypedSelector(selectFormData);
-  const dispatch = useTypedDispatch();
-  const history = useHistory();
+  const initialValues = useStore(selectFormData);
+  const { isNew, event, publishNewEvent, publishEventUpdate } = useStore((state) => state.editor);
+  const navigate = useNavigate();
   const { t } = useTranslation();
-
-  const isNew = useTypedSelector((state) => state.editor.isNew);
-  const eventId = useTypedSelector((state) => state.editor.event?.id);
 
   const onSubmit = useEvent(async (data: EditorEvent, form: FormApi<EditorEvent>) => {
     try {
       let saved;
       if (isNew) {
-        saved = await dispatch(publishNewEvent(data));
-        history.push(paths.adminEditEvent(saved.id));
+        saved = await publishNewEvent(data);
+        navigate(paths.adminEditEvent(saved.id));
         toast.success(t("editor.status.createSuccess"), { autoClose: 2000 });
       } else {
-        saved = await dispatch(publishEventUpdate(eventId!, data));
+        saved = await publishEventUpdate(event!.id, data);
         if (saved) {
           toast.success(t("editor.status.saveSuccess"), { autoClose: 2000 });
         }