+++ packages/ilmomasiina-frontend/src/routes/Editor/components/LanguageVersions.tsx
@@ -7,9 +7,8 @@ import { useTranslation } from "react-i18next";
 
 import { AdminEventLanguage } from "@tietokilta/ilmomasiina-models";
 import { knownLanguages } from "../../../i18n";
-import { languageSelected } from "../../../modules/editor/actions";
 import { EditorEvent } from "../../../modules/editor/types";
-import { useTypedDispatch, useTypedSelector } from "../../../store/reducers";
+import useStore from "../../../modules/store";
 import useEvent from "../../../utils/useEvent";
 
 type VersionProps = { language: string };
@@ -23,15 +22,14 @@ const LanguageVersion = ({ language }: VersionProps) => {
     input: { value: languages },
   } = useField<EditorEvent["languages"]>("languages");
   const { change, getState } = useForm<EditorEvent>();
-  const selectedLanguage = useTypedSelector((state) => state.editor.selectedLanguage);
-  const dispatch = useTypedDispatch();
+  const { selectedLanguage, languageSelected } = useStore((state) => state.editor);
 
   const isPresent = language in languages;
   const isDefault = language === defaultLanguage;
   const isEditing = selectedLanguage === language;
 
   const edit = useEvent(() => {
-    dispatch(languageSelected(language));
+    languageSelected(language);
   });
 
   const setAsDefault = useEvent(() => {
@@ -76,7 +74,7 @@ const LanguageVersion = ({ language }: VersionProps) => {
     }
     change("defaultLanguage", language);
     // Finally, start editing the new default language.
-    dispatch(languageSelected(language));
+    languageSelected(language);
   });
 
   const add = useEvent(() => {
@@ -100,14 +98,14 @@ const LanguageVersion = ({ language }: VersionProps) => {
       },
     });
     // Finally, start editing the new language.
-    dispatch(languageSelected(language));
+    languageSelected(language);
   });
 
   const remove = useEvent(() => {
     // Remove language from event.
     change("languages", omit(languages, language));
     // Switch to default language if we removed the active one.
-    if (isEditing) dispatch(languageSelected(defaultLanguage));
+    if (isEditing) languageSelected(defaultLanguage);
   });
 
   const labels: string[] = [];