+++ packages/ilmomasiina-frontend/src/routes/Editor/components/SignupsTab.tsx
@@ -5,9 +5,8 @@ import { useTranslation } from "react-i18next";
 
 import { stringifyAnswer } from "@tietokilta/ilmomasiina-client/dist/utils/signupUtils";
 import { AdminEventResponse, SignupStatus } from "@tietokilta/ilmomasiina-models";
-import { deleteSignup, editNewSignup, editSignup, getEvent } from "../../../modules/editor/actions";
 import type { AdminSignupWithQuota } from "../../../modules/editor/types";
-import { useTypedDispatch, useTypedSelector } from "../../../store/reducers";
+import useStore from "../../../modules/store";
 import { useActionDateTimeFormatter } from "../../../utils/dateFormat";
 import useEvent from "../../../utils/useEvent";
 import CheckMembershipsTextArea from "./CheckMembershipsTextArea";
@@ -28,20 +27,20 @@ type SignupProps = {
 };
 
 const SignupRow = ({ position, signup, showQuota }: SignupProps) => {
-  const event = useTypedSelector((state) => state.editor.event)!;
-  const dispatch = useTypedDispatch();
+  const { editSignup, deleteSignup, getEvent } = useStore((state) => state.editor);
+  const event = useStore((state) => state.editor.event!);
   const { t } = useTranslation();
   const actionDateFormat = useActionDateTimeFormatter();
 
   const answersMap = useMemo(() => getAnswersFromSignup(event, signup), [event, signup]);
 
-  const onEdit = useEvent(() => dispatch(editSignup(signup)));
+  const onEdit = useEvent(() => editSignup(signup));
   const onDelete = useEvent(async () => {
     // eslint-disable-next-line no-alert
     const confirmation = window.confirm(t("editor.signups.action.delete.confirm"));
     if (confirmation) {
-      await dispatch(deleteSignup(signup.id!));
-      dispatch(getEvent(event.id));
+      await deleteSignup(signup.id!);
+      getEvent(event.id);
     }
   });
 
@@ -53,14 +52,14 @@ const SignupRow = ({ position, signup, showQuota }: SignupProps) => {
       : null;
 
   return (
-    <tr className={!signup.confirmed ? "text-muted" : ""}>
+    <tr className={!signup.confirmed ? "ilmo--unconfirmed" : ""}>
       <td key="position">{`${position}.`}</td>
       {signup.confirmed && event.nameQuestion && <td key="firstName">{signup.firstName}</td>}
       {signup.confirmed && event.nameQuestion && <td key="lastName">{signup.lastName}</td>}
       {signup.confirmed && event.emailQuestion && <td key="email">{signup.email}</td>}
       {signup.confirmed && event.emailQuestion && <td key="membership">{signup.isMember && "\u2705"}</td>}
       {!signup.confirmed && nameEmailCols && (
-        <td colSpan={nameEmailCols} className="font-italic">
+        <td colSpan={nameEmailCols} className="fst-italic">
           {t("editor.signups.unconfirmed")}
         </td>
       )}
@@ -75,7 +74,7 @@ const SignupRow = ({ position, signup, showQuota }: SignupProps) => {
         <Button type="button" variant="primary" size="sm" onClick={onEdit}>
           {t("editor.signups.action.edit")}
         </Button>
-        <Button type="button" variant="danger" size="sm" onClick={onDelete}>
+        <Button type="button" variant="danger" size="sm" onClick={onDelete} className="ms-1">
           {t("editor.signups.action.delete")}
         </Button>
       </td>
@@ -125,8 +124,7 @@ const csvOptions: CSVOptions = { delimiter: "\t" };
 const SignupsTab = () => {
   const [emails, setEmails] = useState<string[]>([]);
 
-  const event = useTypedSelector((state) => state.editor.event);
-  const dispatch = useTypedDispatch();
+  const { event, editNewSignup } = useStore((state) => state.editor);
 
   const signups = useMemo(() => event && getSignupsForAdminList(event, emails), [event, emails]);
   const signupsByQuota = useMemo(() => event && getSignupsByQuotaForAdminList(event, emails), [event, emails]);
@@ -142,7 +140,7 @@ const SignupsTab = () => {
     i18n: { language },
   } = useTranslation();
 
-  const createSignup = useEvent(() => dispatch(editNewSignup({ language })));
+  const createSignup = useEvent(() => editNewSignup(language));
 
   if (!event || !event.quotas.length) {
     return <p>{t("editor.signups.noQuotas")}</p>;
@@ -160,17 +158,18 @@ const SignupsTab = () => {
           checked={grouped}
           onChange={onGroupedChange}
         />
-        <div className="flex-grow-1" />
-        <Button variant="primary" onClick={createSignup}>
-          {t("editor.signups.action.create")}
-        </Button>
-        <CSVLink
-          data={csvSignups!}
-          csvOptions={csvOptions}
-          download={t("editor.signups.download.filename", { event: event.title })}
-        >
-          {t("editor.signups.download")}
-        </CSVLink>
+        <div className="ilmo--title-nav-buttons">
+          <Button variant="primary" onClick={createSignup}>
+            {t("editor.signups.action.create")}
+          </Button>
+          <CSVLink
+            data={csvSignups!}
+            csvOptions={csvOptions}
+            download={t("editor.signups.download.filename", { event: event.title })}
+          >
+            {t("editor.signups.download")}
+          </CSVLink>
+        </div>
       </nav>
       {/* eslint-disable-next-line no-nested-ternary */}
       {!signups?.length ? (